+++ includes/data-stores/class-wc-shipping-zone-data-store.php
@@ -25,7 +25,8 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 	public function create( &$zone ) {
 		global $wpdb;
 		$wpdb->insert(
-			$wpdb->prefix . 'woocommerce_shipping_zones', array(
+			$wpdb->prefix . 'woocommerce_shipping_zones',
+			array(
 				'zone_name'  => $zone->get_zone_name(),
 				'zone_order' => $zone->get_zone_order(),
 			)
@@ -48,10 +49,12 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 		global $wpdb;
 		if ( $zone->get_id() ) {
 			$wpdb->update(
-				$wpdb->prefix . 'woocommerce_shipping_zones', array(
+				$wpdb->prefix . 'woocommerce_shipping_zones',
+				array(
 					'zone_name'  => $zone->get_zone_name(),
 					'zone_order' => $zone->get_zone_order(),
-				), array( 'zone_id' => $zone->get_id() )
+				),
+				array( 'zone_id' => $zone->get_id() )
 			);
 		}
 		$zone->save_meta_data();
@@ -109,17 +112,30 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 	 * @return void
 	 */
 	public function delete( &$zone, $args = array() ) {
-		if ( $zone->get_id() ) {
+		$zone_id = $zone->get_id();
+
+		if ( $zone_id ) {
 			global $wpdb;
-			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zone_methods', array( 'zone_id' => $zone->get_id() ) );
-			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zone_locations', array( 'zone_id' => $zone->get_id() ) );
-			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zones', array( 'zone_id' => $zone->get_id() ) );
-			WC_Cache_Helper::incr_cache_prefix( 'shipping_zones' );
-			$id = $zone->get_id();
+
+			// Delete methods and their settings.
+			$methods = $this->get_methods( $zone_id, false );
+
+			if ( $methods ) {
+				foreach ( $methods as $method ) {
+					$this->delete_method( $method->instance_id );
+				}
+			}
+
+			// Delete zone.
+			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zone_locations', array( 'zone_id' => $zone_id ) );
+			$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zones', array( 'zone_id' => $zone_id ) );
+
 			$zone->set_id( null );
+
 			WC_Cache_Helper::incr_cache_prefix( 'shipping_zones' );
 			WC_Cache_Helper::get_transient_version( 'shipping', true );
-			do_action( 'woocommerce_delete_shipping_zone', $id );
+
+			do_action( 'woocommerce_delete_shipping_zone', $zone_id );
 		}
 	}
 
@@ -140,7 +156,7 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 			$raw_methods_sql = "SELECT method_id, method_order, instance_id, is_enabled FROM {$wpdb->prefix}woocommerce_shipping_zone_methods WHERE zone_id = %d";
 		}
 
-		return $wpdb->get_results( $wpdb->prepare( $raw_methods_sql, $zone_id ) ); // phpcs:ignore WordPress.WP.PreparedSQL.NotPrepared
+		return $wpdb->get_results( $wpdb->prepare( $raw_methods_sql, $zone_id ) ); // phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared
 	}
 
 	/**
@@ -190,7 +206,17 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 	 */
 	public function delete_method( $instance_id ) {
 		global $wpdb;
+
+		$method = $this->get_method( $instance_id );
+
+		if ( ! $method ) {
+			return;
+		}
+
+		delete_option( 'woocommerce_' . $method->method_id . '_' . $instance_id . '_settings' );
+
 		$wpdb->delete( $wpdb->prefix . 'woocommerce_shipping_zone_methods', array( 'instance_id' => $instance_id ) );
+
 		do_action( 'woocommerce_delete_shipping_zone_method', $instance_id );
 	}
 
@@ -245,7 +271,7 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 		return $wpdb->get_var(
 			"SELECT zones.zone_id FROM {$wpdb->prefix}woocommerce_shipping_zones as zones
 			LEFT OUTER JOIN {$wpdb->prefix}woocommerce_shipping_zone_locations as locations ON zones.zone_id = locations.zone_id AND location_type != 'postcode'
-			WHERE " . implode( ' ', $criteria ) // phpcs:ignore WordPress.WP.PreparedSQL.NotPrepared
+			WHERE " . implode( ' ', $criteria ) // phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared
 			. ' ORDER BY zone_order ASC, zone_id ASC LIMIT 1'
 		);
 	}
@@ -317,7 +343,8 @@ class WC_Shipping_Zone_Data_Store extends WC_Data_Store_WP implements WC_Shippin
 
 		foreach ( $zone->get_zone_locations( 'edit' ) as $location ) {
 			$wpdb->insert(
-				$wpdb->prefix . 'woocommerce_shipping_zone_locations', array(
+				$wpdb->prefix . 'woocommerce_shipping_zone_locations',
+				array(
 					'zone_id'       => $zone->get_id(),
 					'location_code' => $location->code,
 					'location_type' => $location->type,
